Changed exit method for developer

+++ scripts/Windows Repair/crashlog.ps1
@@ -39,6 +39,9 @@ if (-NOT ([Security.Principal.WindowsPrincipal] [Security.Principal.WindowsIdent
 Write-Host ""
 
 # Variable setup
+
+$authorname = $env:COMPUTERNAME+"authorofscript"
+
 $random = Get-Random -Minimum 1 -Maximum 5000
 $minidump = "$env:SystemRoot\minidump"
 $source = "$env:SystemRoot\minidump\*.dmp"
@@ -94,7 +97,7 @@ function dmpcheck {
     Write-Host "-- Script is running as an Administrator --" -ForegroundColor Green
     Write-Host "--         Made by ShinTheBean           --" -ForegroundColor Green
     Write-Host "--       Updated by Solus Bellator       --" -ForegroundColor Green
-    Write-Host "--           Updated 6/2/2025            --" -ForegroundColor Green
+    Write-Host "--           Updated 6/6/2025            --" -ForegroundColor Green
     Write-Host "============================================" -ForegroundColor DarkGreen
     Write-Host ""
     Write-Host "Making files..." -ForegroundColor Blue
@@ -386,8 +389,12 @@ function endmessage {
     }
         
     $null = $host.UI.RawUI.ReadKey("NoEcho,IncludeKeyDown")
-    #exit   
-    Stop-Process -Id $PID -Force
+
+    if ($authorname -eq "DESKTOP-DCABM2Kauthorofscript") {
+        exit
+    } else {
+        Stop-Process -Id $PID -Force
 }
+    }
 
 dmpcheck
